node-support: bootloader: Use chain-dependent gas wallet URL
Also refactors the bootloader into a few separate files and internalizes
the funds manager API dependency to prevent a cycle.
