+++ node-support/bootloader/Cargo.toml
@@ -14,7 +14,8 @@ tokio = { workspace = true, features = ["full"] }
 
 # === Workspace Dependencies === #
 config = { workspace = true }
-funds-manager-api = { git = "https://github.com/renegade-fi/relayer-extensions.git", rev = "25bfcce" }
+common = { workspace = true, features = ["hmac"] }
+external-api = { workspace = true, features = ["auth"] }
 util = { workspace = true }
 
 # === Misc Dependencies === #
@@ -22,6 +23,7 @@ base64 = "0.22"
 hex = "0.4"
 libp2p = { workspace = true }
 reqwest = { version = "0.11", features = ["json"] }
+serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 tracing = { workspace = true }
 toml = "0.8"
